Renaming files
interface/gemm.hpp -> interface/gemm_interface.hpp
interface/trsm.hpp -> interface/trsm_interface.hpp
