+++ src/interface/blas3/trsm.cpp.in
@@ -27,7 +27,7 @@
 #include "operations/blas_constants.hpp"
 #include "views/view_sycl.hpp"
 #include "policy/sycl_policy_handler.hpp"
-#include "interface/trsm.hpp"
+#include "interface/trsm_interface.hpp"
 #include "operations/blas3/trsm.hpp"
 
 namespace blas {
